chore: reviews returns

+++ registry/network/helpers.go
@@ -16,8 +16,8 @@ import (
 //
 // Returns:
 //
-//	int - 4 for IPv4, 6 for IPv6.
-func (nr *NetworkRegistry) determineIPVersion(ip net.IP) int {
+//	byte - 4 for IPv4, 6 for IPv6.
+func (nr *NetworkRegistry) determineIPVersion(ip net.IP) uint8 {
 	// Check if it's an IPv4 address
 	if len(ip) == net.IPv4len || (len(ip) == net.IPv6len && ip.To4() != nil) {
 		return 4
